requirements.txt
pyyaml~=6.0
boto3>=1.34.20,<2.0
click~=8.0
pydantic>=2.1.0,<3.0
rich>=13.7.0,<14.0
jinja2>=3.1.3,<4.0
jsonpath-ng>=1.6.1,<2.0
SQLAlchemy>=2.0.0,<3.0
alembic>=1.13.1,<2.0